requirements.txt
# Core dependencies for 2SEARX2COOL integrated version
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
flask-jwt-extended>=4.5.0
flask-migrate>=4.0.0
requests>=2.31.0
redis>=5.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
eventlet>=0.33.3
sqlalchemy>=2.0.0

# Music search engine dependencies
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.24.0
spotipy>=2.23.0

# Additional utilities
python-dateutil>=2.8.0
urllib3>=2.0.0